feat: Add modal for selecting esteira in sidebar menu
This commit adds a modal to the sidebar menu component for selecting an "esteira". The modal is triggered when the "Selecionar Esteira" button is clicked. It contains a dropdown menu to choose from a list of available esteiras. Upon selecting an esteira, the user is able to proceed with the selected esteira. This enhancement improves the user experience by providing a more intuitive way to switch between esteiras.