feat: Add HasuraMapperBase, PromptJiraActivityMapper, PromptJiraActivityDto, PromptJiraActivityEntity, and PromptJiraActivityRepository
This commit adds the following files to the repository:
- HasuraMapperBase.java
- PromptJiraActivityMapper.java
- PromptJiraActivityDto.java
- PromptJiraActivityEntity.java
- PromptJiraActivityRepository.java

These files are part of the dashboardMapper, dashboardDto, and dashboardEntity packages in the dash module. They introduce new interfaces and classes related to mapping and data transfer objects for the dashboard functionality. This addition enhances the overall functionality and structure of the application.